+++ kotori/daq/strategy/lan.py
@@ -3,7 +3,7 @@
 import re
 
 from kotori.daq.strategy import StrategyBase
-from kotori.util.common import SmartBunch
+from kotori.util.common import SmartMunch
 
 
 class LanStrategy(StrategyBase):
@@ -34,7 +34,7 @@ class LanStrategy(StrategyBase):
         # decode the topic
         m = self.matcher.match(topic)
         if m:
-            address = SmartBunch(m.groupdict())
+            address = SmartMunch(m.groupdict())
         else:
             address = {}
 
@@ -63,7 +63,7 @@ class LanStrategy(StrategyBase):
         table_suffix = self.get_table_suffix(topology, message_type)
 
         # Use topology information as blueprint for storage address.
-        storage = SmartBunch(topology)
+        storage = SmartMunch(topology)
 
         # Format and sanitize all input parameters used for database addressing.
         # Todo: Investigate using tags additionally to / instead of only "storage.measurement".